requirements.txt
bs4>=0.0.1
beautifulsoup4>=4.11.1
requests>=2.28.1
Pyrogram>=2.0.41
python-dotenv>=0.20.0
TgCrypto>=1.2.3